rolls being added to history
have to preload history to make sure it exists
then figure out best displays thereof
history depends on name too
there's this pattern of a couple of global objects that
look like plump things but there's just one
but that is useful because they're observable.
this really is redux redux.